pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-base"
version = "0.5.0"
description = "Python library for building production-ready AI agents. Async-first with streaming, tool execution, state persistence, and multimodal support."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.75.0",
    "asyncpg>=0.31.0",
    "blake3>=1.0.8",
    "boto3>=1.40.46",
    "litellm>=1.80.11",
    "pillow>=12.1.1",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent_base"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.workspace]
members = [
    "demos/fastapi_server",
]
